Single Payment Prediction Challenge
The forecasted estimated cost of care in 2018 is predicted to be $19850.10. The data was loaded from the amazon data warehouse text files, and the total cost of care, average cost of care and population was calculated for each year in the dataset. Next, to aid in the analysis, the average population change, average change in total cost of care, and the average change in average costs was calculated. The average change in total cost was added to the most recent year (2016) in order to receive an estimated total cost for 2018. The average change in population was also added to the most recent year’s population (2016), and the resulting estimated total cost of 2018 was divided by the estimated population of 2018 in order to get the average cost of care for 2018. The code can be found below.