Bank Credit Card Customer Churn Using R
The banking industry has long been in the forefront of analytics. Analytics has allowed banks and other companies alike to obtain a competitive advantage through utilization of their customer data. By identifying trends within customer data, banks are better able to understand and predict their customers habits. A particular customer habit that directly affects the success of a business is predicting which customers may be at risk of leaving the banks services. This action is referred to as churn. Bank customer data was retrieved from Saint Joseph’s University’s data mining course, is authentic bank credit card customer information, and will be utilized in order to model the probability of customers staying with a service or leaving a service. The data that was collected and considered during this analysis was churn flag, customer ID, average daily balance, interest paid, cash advances, balance transferred, marital status, occupation group, age of account in months, age, billing cycle, gender, customer value, and credit limit. The dataset contains 245,465 observations and has 14 columns of data. The objective of this project is to use the data to build a model that would predict potential customer churn with high accuracy. Different models will be executed in R statistical programming language and the model with the lowest misclassification rate will become to model of choice to predict customer churn. The models of interest will be logistic regression, decision tree, neural network model due to the necessity to classify new and existing customers as potential churn candidates. The description of the data can be found below.
• Churn Flag – (Target Variable) Represented by 1 for churn and 0 for no churn.
• Customer ID – The identification number of a particular customer of the bank.
• Average Daily Balance – The moving average of the daily balance of the customer’s credit
card at the bank.
• Interest Paid – The interest paid to the bank by the customer the total length of the account life.
• Cash Advances – Cash provided to the customer by the bank in loan form.
• Balance Transferred – The amount of debt transferred to another account.
• Marital Status – The marital status of the customer.
• Occupation Group – The industry in which the customer works in.
• Age of Account – The age of the account in months.
• Age Group – The age of the customer.
• Bill Cycle – What day of the month the customer’s bill cycle stops.
• Gender – Whether customer was a male or female.
• Customer Value – The banks assessed value of the customer.